feat(backoffice): master-detail editor for ticket sections
Replace the flat sortable section-card list with a master-detail editor:
a left rail of section rows (thumbnail, label, product count and
targeting hint, reorderable) and a right pane editing the selected
section content-first. Centralize the whole sections UX into a shared
SectionsEditor, so ticket-select, ticket-card and housing configs drop
their duplicated list handling. The ProductSection model and onChange
contract are unchanged.
