+++ backoffice/src/components/ticketing-step-builder/template-configs/TicketSelectConfig.tsx
@@ -1,35 +1,20 @@
-import {
-  closestCenter,
-  DndContext,
-  type DragEndEvent,
-  PointerSensor,
-  useSensor,
-  useSensors,
-} from "@dnd-kit/core"
-import {
-  arrayMove,
-  SortableContext,
-  verticalListSortingStrategy,
-} from "@dnd-kit/sortable"
 import { useQuery } from "@tanstack/react-query"
-import { Check, Plus } from "lucide-react"
+import { Check } from "lucide-react"
 
 import {
   AttendeeCategoriesService,
   FormFieldsService,
   ProductsService,
 } from "@/client"
-import { Button } from "@/components/ui/button"
 import { Label } from "@/components/ui/label"
 import { Separator } from "@/components/ui/separator"
 import { cn } from "@/lib/utils"
+import { SectionsEditor } from "./sections/SectionsEditor"
 import {
   type ProductSection,
   parseConfigSections,
-  SortableSectionCard,
-  toKey,
   type VisibilityFormFieldOption,
-} from "./SortableSectionCard"
+} from "./sections/sectionTypes"
 import type { TemplateConfigProps } from "./types"
 
 // Only fields with a fixed set of answers can drive section visibility.
@@ -217,50 +202,8 @@ export function TicketSelectConfig({
       return { id: c.id, key: c.key, label }
     })
 
-  const sensors = useSensors(
-    useSensor(PointerSensor, { activationConstraint: { distance: 5 } }),
-  )
-
   const updateSections = (sections: TicketSelectSection[]) => {
     onChange({ ...config, sections })
-  }
-
-  const handleSectionUpdate = (
-    key: string,
-    updates: Partial<TicketSelectSection>,
-  ) => {
-    updateSections(
-      parsed.sections.map((s) => (s.key === key ? { ...s, ...updates } : s)),
-    )
-  }
-
-  const handleSectionDelete = (key: string) => {
-    updateSections(parsed.sections.filter((s) => s.key !== key))
-  }
-
-  const handleDragEnd = (event: DragEndEvent) => {
-    const { active, over } = event
-    if (!over || active.id === over.id) return
-
-    const oldIndex = parsed.sections.findIndex((s) => s.key === active.id)
-    const newIndex = parsed.sections.findIndex((s) => s.key === over.id)
-    if (oldIndex === -1 || newIndex === -1) return
-
-    const reordered = arrayMove(parsed.sections, oldIndex, newIndex).map(
-      (s, i) => ({ ...s, order: i }),
-    )
-    updateSections(reordered)
-  }
-
-  const handleAddSection = () => {
-    const newLabel = `Section ${parsed.sections.length + 1}`
-    const newSection: TicketSelectSection = {
-      key: `${toKey(newLabel)}-${Date.now()}`,
-      label: newLabel,
-      order: parsed.sections.length,
-      product_ids: [],
-    }
-    updateSections([...parsed.sections, newSection])
   }
 
   return (
@@ -322,54 +265,15 @@ export function TicketSelectConfig({
 
       <Separator />
 
-      {/* Sections */}
-      <div className="flex items-center justify-between">
-        <div>
-          <Label className="text-sm font-medium">Sections</Label>
-          <p className="text-xs text-muted-foreground">
-            Configure how products are grouped in the checkout
-          </p>
-        </div>
-        <Button variant="outline" size="sm" onClick={handleAddSection}>
-          <Plus className="h-3.5 w-3.5 mr-1" />
-          Add Section
-        </Button>
-      </div>
-
-      {parsed.sections.length === 0 ? (
-        <div className="rounded-lg border border-dashed p-6 text-center text-sm text-muted-foreground">
-          No sections configured. Products will be displayed as a flat list.
-        </div>
-      ) : (
-        <DndContext
-          sensors={sensors}
-          collisionDetection={closestCenter}
-          onDragEnd={handleDragEnd}
-        >
-          <SortableContext
-            items={parsed.sections.map((s) => s.key)}
-            strategy={verticalListSortingStrategy}
-          >
-            <div className="flex flex-col gap-2">
-              {parsed.sections
-                .sort((a, b) => a.order - b.order)
-                .map((section) => (
-                  <SortableSectionCard
-                    key={section.key}
-                    section={section}
-                    onUpdate={handleSectionUpdate}
-                    onDelete={handleSectionDelete}
-                    products={products}
-                    showMediaFields={false}
-                    showAttendeeCategories={true}
-                    attendeeCategories={attendeeCategories}
-                    visibilityFormFields={visibilityFormFields}
-                  />
-                ))}
-            </div>
-          </SortableContext>
-        </DndContext>
-      )}
+      <SectionsEditor
+        sections={parsed.sections}
+        onChange={updateSections}
+        products={products}
+        showMediaFields={false}
+        showAttendeeCategories={true}
+        attendeeCategories={attendeeCategories}
+        visibilityFormFields={visibilityFormFields}
+      />
     </div>
   )
 }
